backend/requirements.txt
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
django-cors-headers==4.9.0
python-dateutil==2.9.0
icalendar==6.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
